% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundaryPlot.R
\name{BoundaryPlot}
\alias{BoundaryPlot}
\title{Title Boundary plot}
\usage{
BoundaryPlot(
  TumorSTn = TumorSTn,
  TumorST = TumorST,
  OutDir = OutDir,
  Sample = Sample
)
}
\arguments{
\item{TumorSTn}{A subset Seurat object of ST data with defined malignant (Mal) spots, boundary (Bdy) spots, and non-malignant (nMal) spots}

\item{TumorST}{A Seurat object with morphological adjusted expression matrix and determined clusters}

\item{OutDir}{Path to file save figures and processed data}

\item{Sample}{Name of your sample}
}
\value{
A Seurat object with defined malignant (Mal) spots, boundary (Bdy) spots, and non-malignant (nMal) spots
}
\description{
Visualize boundary define results of ST data
}
\examples{
\dontrun{
  Sample <- "CRC1"
  OutDir <- "YourPath/TumorBoundary/Fig/1.BoundaryDefine/CRC1/"
  TumorST <- readr::read_rds("YourPath/TumorBoundary/1.BoundaryDefine/CRC1/TumorSTClustered.rds.gz")
  TumorST <- STCNVScore(TumorST = TumorST, assay = "Spatial",
                        OutDir = OutDir, Sample = Sample)
  MalLabel <- c(1, 2)
  TumorSTn <- BoundaryDefine(TumorST = TumorST, MalLabel = MalLabel,
                             OutDir = OutDir, Sample = Sample)
  TumorST <- BoundaryPlot(TumorSTn = TumorSTn, TumorST = TumorST,
                          OutDir = OutDir, Sample = Sample)
}

}
