% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{ClusterUpdate}
\alias{ClusterUpdate}
\title{Title update cluster data frame}
\usage{
ClusterUpdate(
  MalCellIDN = MalCellIDN,
  position = position,
  df_j = df_j,
  UMAPembeddings = UMAPembeddings,
  NormalCellID = NormalCellID,
  BdyCellID = BdyCellID,
  MalCellID = MalCellID,
  x = 100
)
}
\arguments{
\item{MalCellIDN}{barcodes of newlydefined malignant (Mal) spots}

\item{position}{<data.frame> spatial spots information(spot id, row, cols ,imagerow, imagecol)}

\item{df_j}{a list contain spatial spots and thier neighbors}

\item{UMAPembeddings}{<data.frame> spatial spots' umap reduction coordinates}

\item{NormalCellID}{barcodes of normal spots}

\item{BdyCellID}{barcodes of boundary (Bdy) spots}

\item{MalCellID}{barcodes of all malignant spots}

\item{x}{repeat times}
}
\value{
<data.frame> neighbors of new defined malignant spots' ids and their define types
}
\description{
Title update cluster data frame
}
