% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconBarplot.R
\name{DeconBarplot}
\alias{DeconBarplot}
\title{Title Bar plot of deconvolution result}
\usage{
DeconBarplot(DeconData, TumorST, plot_col)
}
\arguments{
\item{DeconData}{A data frame of deconvolution result of each spatial spot}

\item{TumorST}{A Seurat object with defined malignant (Mal) spots, boundary (Bdy) spots, and non-malignnat (nMal) spots}

\item{plot_col}{Names of cell types to plot}
}
\value{
A ggplot object
}
\description{
Bar plot of deconvolution result in location groups
}
\examples{
\dontrun{
  TumorST <- readr::read_rds("YourPath/TumorSTBoundaryDefine.rds.gz")
  DeconData <- oepnxlsx::read.xlsx(system.file("extdata/DeconData.xlsx", package = "Cottrazm"))
  plot_col <- colnames(DeconData)[2:ncol(DeconData)]
  DeconBarplot(DeconData = DeconData, TumorST = TumorST, plot_col = plot_col)
}

}
