% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconPieplot.R
\name{DeconPieplot}
\alias{DeconPieplot}
\title{Title Pie plot of deconvolution result}
\usage{
DeconPieplot(
  DeconData = DeconData,
  TumorST = TumorST,
  plot_col = plot_col,
  img_path = img_path,
  pie_scale = pie_scale,
  scatterpie_alpha = scatterpie_alpha,
  border_color = border_color
)
}
\arguments{
\item{DeconData}{A data frame of deconvolution result of each spatial spot}

\item{TumorST}{A Seurat object with defined malignant (Mal) spots, boundary (Bdy) spots, and non-malignnat (nMal) spots}

\item{plot_col}{Names of cell types to plot}

\item{img_path}{Path to tissue_lowers_image.jpg in Spaceranger output file}

\item{pie_scale}{Size of spot in pie plot}

\item{scatterpie_alpha}{Transparency of spot in pie plot}

\item{border_color}{Color of border in pie plot}
}
\value{
A ggplot object each pie represented a ST spot, each part of the pie represent the percentage of cell types infiltrated in the ST spot
}
\description{
Pie plot of deconvolution result
}
\examples{
\dontrun{
  TumorST <- readr::read_rds("YourPath/TumorSTBoundaryDefine.rds.gz")
  DeconData <- openxlsx::read.xlsx(system.file("extdata/DeconData.xlsx", package = "Cottrazm"))
  plot_col <- colnames(DeconData)[2:ncol(DeconData)]
  img_path <- system.file("extdata/outs/spatial/tissue_lowres_image.png", package = "Cottrazm")
  DeconPieplot(DeconData = DeconData, TumorST = TumorST, plot_col = plot_col, img_path = img_path,
               pie_scale = 0.4, scatterpie_alpha = 0.8, border_color = "grey")
}

}
