% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_analysis.R
\name{enrich_analysis}
\alias{enrich_analysis}
\title{Title enrich analysis}
\usage{
enrich_analysis(filter_log_expr, enrich_matrix)
}
\arguments{
\item{filter_log_expr}{A matrix of filtered log transformed ST data expression}

\item{enrich_matrix}{A binary matrix features as row, cell types as column}
}
\value{
A matrix of enrich result, cell types as row, spatial spots as column
}
\description{
PAGE enrich analysis
}
\examples{
\dontrun{
  TumorST <- readr::read_rds("YourPath/TumorSTBoundaryDefine.rds.gz")
  sig_exp <- readr::read_rds("YourPath/sig_exp.rds.gz")
  clustermarkers_list <- readr::read_rds("YourPath/clustermarkers_list.rds.gz")
  # get st and sc filtered expr data
  TumorST <- NormalizeData(TumorST,assay = "Spatial")
  TumorST@meta.data$Decon_topics <- paste(TumorST@meta.data$Location,
                                          TumorST@meta.data$seurat_clusters,sep = "_")
  expr_values = as.matrix(TumorST@assays$Spatial@data) #ST expr log
  nolog_expr = 2^(expr_values)-1 #ST expr nolog
  # filter st and sc feature
  intersect_gene = intersect(rownames(sig_exp), rownames(nolog_expr))
  filter_sig = sig_exp[intersect_gene,]
  filter_log_expr = expr_values[intersect_gene,]
  # get enrich matrix
  enrich_matrix <- get_enrich_matrix(filter_sig,clustermarkers_list)
  enrich_result <- enrich_analysis(filter_log_expr, enrich_matrix)
}

}
