% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEnrichment.R
\name{FeatureEnrichment}
\alias{FeatureEnrichment}
\title{Title Enrichment analysis of features of different locations}
\usage{
FeatureEnrichment(
  DiffGenes,
  cut_off_logFC = 0.25,
  cut_off_pvalue = 0.05,
  Location = c("Bdy", "Mal", "nMal")
)
}
\arguments{
\item{DiffGenes}{A list of data frame contain differential expressed genes, p value, log2FC and adjusted p value in Mal, Bdy and nMal spots}

\item{cut_off_logFC}{Cut off of log2FC, features log2FC > cut_off_logFC will be selected as significant}

\item{cut_off_pvalue}{Cut off of p value, features p value < cut_off_pvalue will be selected as significant}

\item{Location}{Names of location to analyse feature enrichment.}
}
\value{
A tibble contain KEGG and GO enrichment result
}
\description{
Use differential expressed genes of location to define the KEGG and GO pathway these DEG enriched.
}
\examples{
\dontrun{
  TumorST <- readr::read_rds("YourPath/TumorSTBoundaryDefine.rds.gz")
  DiffGenes <- FindDiffGenes(TumorST = TumorST, assay = "Spatial")
  BdyFeatureEnrich <- FeatureEnrichment(DiffGenes = DiffGenes, cut_off_logFC = 0.25,
                                        cut_off_pvalue = 0.05, Location = "Bdy")
}

}
