% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{find_neighbors}
\alias{find_neighbors}
\title{Title find neighbors}
\usage{
find_neighbors(position, radius, method = c("manhattan", "euclidean"))
}
\arguments{
\item{position}{<data.frame> spatial spots information(spot id, row, cols ,imagerow, imagecol)}

\item{radius}{predicted radius of spatial spots}

\item{method}{method to calculate distance}
}
\value{
a list contain of all spatial spots' neighbors
}
\description{
Title find neighbors
}
