% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDiffGenes.R
\name{FindDiffGenes}
\alias{FindDiffGenes}
\title{Title Find differential expressed genes}
\usage{
FindDiffGenes(TumorST = TumorST, assay = c("Morph", "Spatial"))
}
\arguments{
\item{TumorST}{A Seurat object with defined  malignant (Mal) spots, boundary (Bdy) spots, and  non-malignant (nMal) spots}

\item{assay}{The name of assay used to find DEG (Morph: Morphological adjusted gene expression, Spatial: gene expression)}
}
\value{
A list of data frame contain differential expressed genes, p-value, log2FC and adjusted p value in Mal, Bdy, and nMal spots
}
\description{
Find differential expressed genes in malignant (Mal) spots, boundary (Bdy) spots, and  non-malignant (nMal) spots
}
\examples{
\dontrun{
  TumorST <- readr::read_rds("YourPath/TumorSTBoundaryDefine.rds.gz")
  assay <- "Spatial"
  DiffGenes <- FindDiffGenes(TumorST = TumorST, assay = assay)
}

}
