% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrich_matrix.R
\name{get_enrich_matrix}
\alias{get_enrich_matrix}
\title{Title get enrich matrix}
\usage{
get_enrich_matrix(filter_sig, clustermarkers_list)
}
\arguments{
\item{filter_sig}{The sig_exp filtered by features}

\item{clustermarkers_list}{A list of markers of each cell type of scRNAseq dataset}
}
\value{
A binary matrix features as row, cell types as column
}
\description{
Construct binary matrix for PAGE enrichment
}
