% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{nbrs}
\alias{nbrs}
\title{Title get neighbors of spot}
\usage{
nbrs(df_j = df_j, MalSpotIDAdd = MalSpotIDAdd, SpotIDRaw = SpotIDRaw)
}
\arguments{
\item{df_j}{<list> contain spatial spot and its' neighbors}

\item{MalSpotIDAdd}{<character> spatial malignant spots' id to find neighbors}

\item{SpotIDRaw}{<character> spatial spots' id already defined}
}
\value{
<list> contain spatial spot and its' neighbors
}
\description{
get neighbor spots' id of MalSpotIDAdd
}
