% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCNVScore.R
\name{STCNVScore}
\alias{STCNVScore}
\title{Title Score ST data with infercnv result}
\usage{
STCNVScore(
  TumorST = TumorST,
  assay = c("Saptial", "Morph"),
  OutDir = NULL,
  Sample = Sample
)
}
\arguments{
\item{TumorST}{A Seurat object with morphological adjusted expression matrix and determined clusters}

\item{assay}{Name of assay uesed in infercnv}

\item{OutDir}{Path to file save figures and processed data}

\item{Sample}{Name of your sample}
}
\value{
A Seurat object of ST data with CNV sub-clusters and CNV scores
}
\description{
add cnv score to ST Seurat object
}
\examples{
\dontrun{
  TumorST <- readRDS("YourPath/TumorBoundary/1.BoundaryDefine/CRC1/TumorSTClustered.rds.gz")
  OutDir <- "YourPath/TumorBoundary/Fig/1.BoundaryDefine/CRC1/"
  Sample <- "CRC1"
  assay = "Spatial"
  TumorST <- STCNVScore(TumorST = TumorST, assay = assay,
                        OutDir = OutDir, Sample = Sample)
}

}
