% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STPreProcess.R
\name{STPreProcess}
\alias{STPreProcess}
\title{Title ST data preprocess}
\usage{
STPreProcess(InDir = InDir, Sample = Sample, OutDir = NULL)
}
\arguments{
\item{InDir}{Path to file contain Spaceranger out result (at least including 'filtered_feature_bc_matrix','filtered_feature_bc_matrix.h5','spatial')}

\item{Sample}{Name of your sample}

\item{OutDir}{Path to file save figures and processed data}
}
\value{
A Seurat object
}
\description{
read ST raw data and quality control
}
\examples{
\dontrun{
  InDir <- paste(system.file("extdata/outs", package = "Cottrazm"), "/", sep = "")
  Sample <- "CRC1"
  OutDir <- "YourPath/TumorBoundary/Fig/1.BoundaryDefine/CRC1/"
  TumorST <- STPreProcess(InDir = InDir, OutDir = OutDir, Sample = Sample)
}

}
