% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransitionPlot.R
\name{TransitionPlot}
\alias{TransitionPlot}
\title{Title Transition plot}
\usage{
TransitionPlot(
  TumorSTn = TumorSTn,
  TumorST = TumorST,
  OutDir = OutDir,
  Sample = Sample
)
}
\arguments{
\item{TumorSTn}{A subset Seurat object of ST data with defined tumor, transition zone, and normal spots}

\item{TumorST}{A Seurat object with morphological adjusted expression matrix and determined clusters}

\item{OutDir}{Path to file save figures and processed data}

\item{Sample}{Name of your sample}
}
\value{
A Seurat object with defined tumor spots, transition zone spots, and out spots
}
\description{
Visualize transition define results of ST data
}
\examples{

Sample = "CRC1"
OutDir = "YourPath/TumorTransition/Fig/1.TransitionDefine/CRC1/"
infercnv.dend <- read.tree(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations_dendrogram.txt",package = "Cottrazm"))
cnv_table <- read.table(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations.txt",package = "Cottrazm"))
TumorST <- readr::read_rds("YourPath/TumorTransition/1.TransitionDefine/CRC1/TumorSTClustered.rds.gz")
TumorST <- STCNVScore(infercnv.dend = infercnv.dend,cnv_table = cnv_table,TumorST = TumorST,OutDir = OutDir)
TumorLabel = c(1,2)
TumorSTn <- TransitionDefine(TumorST = TumorST, TumorLabel = TumorLabel, OutDir = OutDir, Sample = Sample)
TumorST <- TransitionPlot(TumorSTn = TumorSTn, TumorST = TumorST, OutDir = OutDir,Sample = Sample)

}
