% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoundaryDefine.R
\name{BoundaryDefine}
\alias{BoundaryDefine}
\title{Title Boundary define}
\usage{
BoundaryDefine(
  TumorST = TumorST,
  TumorLabel = NULL,
  OutDir = NULL,
  Sample = Sample
)
}
\arguments{
\item{TumorST}{A Seurat object with morphological adjusted expression matrix and determined clusters}

\item{TumorLabel}{CNV labels with high cnv scores which could be defined as tumor. If NULL, Cottrazm will take the two CNVLabels with the highest average CNVScores.}

\item{OutDir}{Path to file save figures and processed data}

\item{Sample}{Name of your sample}
}
\value{
A subset Seurat object of ST data with defined tumor, boundary/transition zone, and normal spots
}
\description{
Find boudnary/transition zone of tumor in ST data
}
\examples{
TumorST <- readr::read_rds("YourPath/Tumorboundary/1.BoundaryDefine/CRC1/TumorSTClustered.rds.gz")
Sample = "CRC1"
OutDir = "YourPath/Tumorboundary/Fig/1.BoundaryDefine/CRC1/"
infercnv.dend <- read.tree(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations_dendrogram.txt",package = "Cottrazm"))
cnv_table <- read.table(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations.txt",package = "Cottrazm"))
TumorST <- STCNVScore(infercnv.dend = infercnv.dend,cnv_table = cnv_table,TumorST = TumorST,OutDir = OutDir)
TumorLabel = c(1,2)
TumorSTn <- BoundaryDefine(TumorST = TumorST, TumorLabel = TumorLabel, OutDir = OutDir, Sample = Sample)

}
