% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{ClusterUpdate}
\alias{ClusterUpdate}
\title{Title updat cluster dataframe update}
\usage{
ClusterUpdate(
  TumorCellIDN = TumorCellIDN,
  position = position,
  df_j = df_j,
  UMAPembeddings = UMAPembeddings,
  NormalCellID = NormalCellID,
  TransCellID = TransCellID,
  TumorCellID = TumorCellID,
  x = n
)
}
\arguments{
\item{TumorCellIDN}{<character> new defined tumor spots ids}

\item{position}{<data.frame> spatial spots information(spot id, row, cols ,imagerow, imagecol)}

\item{df_j}{<list> contain spaital spot and its' neighbors}

\item{UMAPembeddings}{<data.frame> spatial spots umap reduction coordinates}

\item{NormalCellID}{<character> normal spots ids}

\item{TransCellID}{<character> trans spots ids}

\item{TumorCellID}{<characoter> all tumor spots ids}

\item{x}{<interger> repeat times}
}
\value{
<data.frame> neighbors of new defined tumor spots ids and their define types
}
\description{
Title updat cluster dataframe update
}
