% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{ClusterUpdate}
\alias{ClusterUpdate}
\title{Title update cluster data frame}
\usage{
ClusterUpdate(
  MalCellIDN = MalCellIDN,
  position = position,
  df_j = df_j,
  UMAPembeddings = UMAPembeddings,
  NormalCellID = NormalCellID,
  BdyCellID = BdyCellID,
  MalCellID = MalCellID,
  x = n
)
}
\arguments{
\item{MalCellIDN}{<character> new defined malignant (Mal) spots' ids}

\item{position}{<data.frame> spatial spots information(spot id, row, cols ,imagerow, imagecol)}

\item{df_j}{<list> contain spaital spot and its' neighbors}

\item{UMAPembeddings}{<data.frame> spatial spots' umap reduction coordinates}

\item{NormalCellID}{<character> normal spots' ids}

\item{BdyCellID}{<character> boundary (Bdy) spots' ids}

\item{MalCellID}{<character> all malignant spots' ids}

\item{x}{<integer> repeat times}
}
\value{
<data.frame> neighbors of new defined malignant spots' ids and their define types
}
\description{
Title update cluster data frame
}
