% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_define_utility.R
\name{compute_interspot_distances}
\alias{compute_interspot_distances}
\title{Title compute_interspots_distance}
\usage{
compute_interspot_distances(position, scale.factor = 1.05)
}
\arguments{
\item{position}{a data frame contain spatial spots' information(spot id, row, cols ,imagerow, imagecol)}

\item{scale.factor}{scale factor}
}
\value{
a list contain predicted radius
}
\description{
Title compute_interspots_distance
}
