% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconBarplot.R
\name{DeconBarplot}
\alias{DeconBarplot}
\title{Title Barplot of deconvolution result}
\usage{
DeconBarplot(DeconData = DeconData, TumorST = TumorST, plot_col = plot_col)
}
\arguments{
\item{DeconData}{A data frame of deconvolution result of each spatial spot}

\item{TumorST}{A Seurat object with defined tumor spots, boundary/transition zone spots, and out spots}

\item{plot_col}{Names of cell types to plot}
}
\value{
A ggplot object
}
\description{
Barplot of deconvolution result in boundary define groups
}
\examples{
TumorST <- readr::read_rds("YourPath/Tumorboundary/1.BoundaryDefine/CRC1/TumorSTBoundaryDefine.rds.gz")
DeconData <- read.table(system.file("extdata/DeconData.csv",package = "Cottrazm"),sep = ",",header = T,row.names = "X")
plot_col <- colnames(DeconData)[2:ncol(DeconData)]
DeconBarplot(DeconData = DeconData, TumorST = TumorST, plot_col = plot_col)

}
