% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffVolcanoplot.R
\name{DiffVolcanoplot}
\alias{DiffVolcanoplot}
\title{Title Differential expressed genes volcano plot}
\usage{
DiffVolcanoplot(
  DiffGenes = DiffGenes,
  Location = c("Tumor", "Trans", "OutSpots"),
  cut_off_pvalue = 2,
  cut_off_logFC = 0.25,
  n = 10
)
}
\arguments{
\item{DiffGenes}{A list of data frame contain differential expressed genes and its p value, log2FC and adjusted p value in tumor, boundary/transition zone and out spots}

\item{Location}{Name in boundary defined ST data ("Tumor","Trans","OutSpots")}

\item{cut_off_pvalue}{Cut off of p value, features p value < 10^(-cut_off_pvalue) will be selected as significant}

\item{cut_off_logFC}{Cut off of log2FC, features log2FC > cut_off_logFC will be selected as significant}

\item{n}{Number of features labeled in volcano plot, default is 10}
}
\value{
A ggplot object
}
\description{
Draw a volcano plot of features in boundary defined ST data
}
\examples{
TumorST <- readr::read_rds("YourPath/Tumorboundary/1.BoundaryDefine/CRC1/TumorSTBoundaryDefine.rds.gz")
DiffGenes <- FindDiffGenes(TumorST = TumorST,assay = "Morph")
DiffVolcanoplot(DiffGenes = DiffGenes, Location = "Trans", cut_off_pvalue = 2, cut_off_logFC = 0.25, n = 10)

}
