% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_analysis.R
\name{enrich_analysis}
\alias{enrich_analysis}
\title{Title enrich analysis}
\usage{
enrich_analysis(filter_log_expr, enrich_matrix)
}
\arguments{
\item{filter_log_expr}{A matrix of filtered log transformed ST data expression}

\item{enrich_matrix}{A binary matrix features as row, cell types as column}
}
\value{
A matrix of enrich result, cell types as row, spatial spots as column
}
\description{
PAGE enrich analysis
}
\examples{

}
