% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureEnrichment.R
\name{FeatureEnrichment}
\alias{FeatureEnrichment}
\title{Title Enrichment analysis of features of boundary location}
\usage{
FeatureEnrichment(
  DiffGenes = DiffGenes,
  cut_off_logFC = 0.25,
  cut_off_pvalue = 0.05,
  Location = c("Trans", "Tumor", "OutSpots")
)
}
\arguments{
\item{DiffGenes}{A list of dataframe contain differential expressed genes and its p value, log2FC and adjusted p value in tumor, boundary/transition zone and outspots}

\item{cut_off_logFC}{Cut off of log2FC, features log2FC > cut_off_logFC will be selected as significant}

\item{cut_off_pvalue}{Cut off of p value, features p value < cut_off_pvalue will be selected as significant}

\item{Location}{Names of boundary Location to analyse feature enrichment.}
}
\value{
A tibble data frame contain KEGG and GO enrichment result
}
\description{
Use differentially expressed genes of boundary location to find the KEGG and GO pathway these locations DEG enriched.
}
\examples{
TumorST <- readr::read_rds("YourPath/Tumorboundary/1.BoundaryDefine/CRC1/TumorSTBoundaryDefine.rds.gz")
DiffGenes <- FindDiffGenes(TumorST = TumorST,assay = "Spatial")
TransFeatureEnrich <- FeatureEnrichment(DiffGenes = DiffGenes, cut_off_logFC = 0.25, cut_off_pvalue = 0.05, Location = "Trans")

}
