% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDiffGenes.R
\name{FindDiffGenes}
\alias{FindDiffGenes}
\title{Title Find differential expressed genes}
\usage{
FindDiffGenes(TumorST = TumorST, assay = c("Morph", "Spatial"))
}
\arguments{
\item{TumorST}{A Seurat object with defined tumor spots, boundary/transition zone spots, and out spots}

\item{assay}{The name of assay used to find DEG (Morph: Morphological adjusted gene expression, Spatial: gene expression)}
}
\value{
A list of data frame contain differential expressed genes and its p-value, log2FC and adjusted p value in tumor, boundary/transition zone and out spots
}
\description{
Find differential expressed genes in tumor, boundary/transition, and out spots
}
\examples{
TumorST <- readr::read_rds("YourPath/Tumorboundary/1.BoundaryDefine/CRC1/TumorSTBoundaryDefine.rds.gz")
assay = "Spatial"
DiffGenes <- FindDiffGenes(TumorST = TumorST,assay = assay)

}
