% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_define_utility.R
\name{nbrs}
\alias{nbrs}
\title{Title get neighbors of spot}
\usage{
nbrs(df_j = df_j, TumorSpotIDAdd = TumorSpotIDAdd, SpotIDRaw = SpotIDRaw)
}
\arguments{
\item{df_j}{<list> contain spaital spot and its' neighbors}

\item{TumorSpotIDAdd}{<character> spatial spot id to find neighbors}

\item{SpotIDRaw}{<character> spatial spot id already defined}
}
\value{
<list> contain spatial spot and its' neighbors
}
\description{
get neighbor spots' id of TumorSpotIDAdd
}
