% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialDecon.R
\name{SpatialDecon}
\alias{SpatialDecon}
\title{Title Deconvolution of ST spots expression use dampened weighted least squares (DWLS)}
\usage{
SpatialDecon(
  enrich_matrix = enrich_matrix,
  enrich_result = enrich_result,
  filter_expr = filter_expr,
  filter_sig = filter_sig,
  clustermarkers_list = clustermarkers_list,
  meta_data = meta_data,
  malignant_cluster = malignant_clsuter,
  tissue_cluster = tissue_clsuter,
  stromal_cluster = stromal_cluster
)
}
\arguments{
\item{enrich_matrix}{A binary matrix features as row, cell types as column}

\item{enrich_result}{A matrix of enrich result, cell types as row, spatial spots as column}

\item{filter_expr}{A matrix of filtered ST data expression}

\item{filter_sig}{The sig_exp filtered by features, sig_exp: mean expression of each feature across cell types generated by scRNAseq data}

\item{clustermarkers_list}{A list of markers of each cell types of scRNAseq dataset}

\item{meta_data}{A data frame contain nCount_Spatial,decon_topics,and signature scores of each spot}

\item{malignant_cluster}{The name of malignant cluster in scRNAseq dataset, for example "Tumor", "Tumor cells", "Malignant epithelial", "Malignant epithelial cells", etc.}

\item{tissue_cluster}{The name of adjacent normal tissue cluster in scRNAseq dataset, for example "Epithelial", "Epithelial cells", "Hepatocytes", "Cholangiocytes", etc.}

\item{stromal_cluster}{The name of stromal cluster poorly permeated, for example "Fibroblast", "Fibroblast cells", etc.}
}
\value{
A data frame of deconvolution result of each spatial spot
}
\description{
deconvolution of ST spots to explore the infiltration proportion of maintype clusters in each spot
}
\examples{

}
