% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialRecon.R
\name{SpatialRecon}
\alias{SpatialRecon}
\title{Title Reconstruct of spatial tumor micro environment at sub-spot level}
\usage{
SpatialRecon(
  TumorST = TumorST,
  sig_exp = sig_exp,
  clustermarkers_list = clustermarkers_list,
  DeconData = DeconData,
  Location = NULL
)
}
\arguments{
\item{TumorST}{A seurat object with defined malignant (Mal) spots, boundary (Bdy) spots, and non-malignant (nMal) spots}

\item{sig_exp}{Mean expression of each feature across cell types generated by scRNAseq data}

\item{clustermarkers_list}{A list of markers of each cell type  of scRNAseq dataset}

\item{DeconData}{A data frame of deconvolution result of each spatial spot}

\item{Location}{Name of location to reconstruct TME.  More spots selected, reconstruct will take longer time.}
}
\value{
A reconstructed Seurat object of ST tumor data
}
\description{
Reconstruct spatial tumor micro environment at sub-spot level using spot deconvolution result and scRNAseq data
}
\examples{
TumorST <- readr::read_rds("YourPath/TumorBoundary/1.BoundaryDefine/CRC1/TumorSTBoundaryDefine.rds.gz")
sig_exp <- readr::read_rds("YourPath/sig_exp.rds.gz")
clustermarkers_list <- readr::read_rds("YourPath/clustermarkers_list.rds.gz")
DeconData <- oepnxlsx::read.xlsx(system.file("extdata/DeconData.xlsx", package = "Cottrazm"))
TumorSTSub <- SpatialRecon(TumorST = TumorST, sig_exp = sig_exp, clustermarkers_list = clustermarkers_list, DeconData = DeconData, Location = "Bdy")

}
