% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STCNVScore.R
\name{STCNVScore}
\alias{STCNVScore}
\title{Title Score ST data with infercnv result}
\usage{
STCNVScore(
  infercnv.dend = infercnv.dend,
  cnv_table = cnv_table,
  TumorST = TumorST,
  OutDir = NULL,
  Sample = Sample
)
}
\arguments{
\item{infercnv.dend}{InferCNV random tree subclustered result of ST data}

\item{cnv_table}{CNV scores of observations in ST data}

\item{TumorST}{A Seurat object with morphological adjusted expression matrix and determined clusters}

\item{OutDir}{Path to file save figures and processed data}

\item{Sample}{Name of your sample}
}
\value{
A Seurat object of ST data with CNV subclusters and CNV scroes
}
\description{
add cnv score to ST seurat object
}
\examples{
infercnv.dend <- read.tree(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations_dendrogram.txt",package = "Cottrazm"))
cnv_table <- read.table(file = system.file("extdata/17_HMM_predHMMi6.rand_trees.hmm_mode-subclusters.observations.txt",package = "Cottrazm"))
TumorST <- readr::read_rds("YourPath/TumorTransition/1.TransitionDefine/CRC1/TumorSTClustered.rds.gz")
OutDir = "YourPath/TumorTransition/Fig/1.TransitionDefine/CRC1/"
Sample = "CRC1"
TumorST <- STCNVScore(infercnv.dend = infercnv.dend,cnv_table = cnv_table,TumorST = TumorST,OutDir = OutDir,Sample = Sample)
}
