% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModiCluster.R
\name{STModiCluster}
\alias{STModiCluster}
\title{Title Morphological adjusted cluster determination}
\usage{
STModiCluster(
  InDir = InDir,
  Sample = Sample,
  OutDir = NULL,
  TumorST = TumorST,
  res = 1.5
)
}
\arguments{
\item{InDir}{Path to file contain Spaceranger out result}

\item{Sample}{Name of your sample}

\item{OutDir}{Path to file save figures and processed data}

\item{TumorST}{A Seurat object}

\item{res}{Value of the resolution parameter}
}
\value{
A Seurat object with morphological adjusted expression matrix and determined clusters
}
\description{
Grep HE staining information as Morphological matrix to adjusted gene expression matrix and use adjusted assay to cluster ST data
}
\examples{
InDir <- paste(system.file("extdata/outs", package = "Cottrazm"), "/", sep = "")
Sample <- "CRC1"
OutDir <- "YourPath/TumorBoundary/Fig/1.BoundaryDefine/CRC1/"
TumorST <- STPreProcess(InDir = InDir, OutDir = OutDir, Sample = Sample)
TumorST <- STModiCluster(InDir = InDir, Sample = Sample, OutDir = OutDir, TumorST = TumorST, res = 1.5)

}
